Use maybe-async-cfg to combine sync and async implementations
