+++ Cargo.toml
@@ -28,7 +28,9 @@ async = ["embedded-hal-async"]
 [dependencies]
 embedded-hal = "0.2.5"
 embedded-hal-async = { version = "0.1.0-alpha.1", optional = true }
+maybe-async-cfg = "0.2.1"
 nb = { version = "1", optional = true }
+trait-set = "0.3.0"
 
 [dev-dependencies]
 embedded-hal-mock = "0.7"
