--- src/async/gesture/mod.rs
@@ -1,2 +1,0 @@
-mod config;
-mod reading;
