--- src/async/mod.rs
@@ -1,5 +1,0 @@
-mod config;
-mod gesture;
-mod light;
-mod proximity;
-mod reading;
