+++ src/lib.rs
@@ -180,13 +180,17 @@
 #![deny(missing_docs, unsafe_code)]
 #![no_std]
 
+use trait_set::trait_set;
+#[cfg(feature = "nb")]
 extern crate embedded_hal as hal;
 #[cfg(feature = "nb")]
-use crate::hal::blocking::i2c;
+use crate::hal::blocking::i2c as i2c;
 #[cfg(feature = "nb")]
 extern crate nb;
 #[cfg(feature = "async")]
-use embedded_hal_async::i2c as async_i2c;
+extern crate embedded_hal_async as hal_async;
+#[cfg(feature = "async")]
+use crate::hal_async::i2c as i2c_async;
 
 /// All possible errors in this crate
 #[derive(Debug)]
@@ -371,6 +375,7 @@ mod register {
 }
 
 /// APDS9960 device driver.
+#[maybe_async_cfg::maybe(sync(feature="nb", keep_self), async(feature="async"))]
 #[derive(Debug, Default)]
 pub struct Apds9960<I2C> {
     /// The concrete I²C device implementation.
@@ -382,10 +387,28 @@ pub struct Apds9960<I2C> {
     gconfig4: register::GConfig4,
 }
 
+// This can go away when we upgrade to embedded-hal 1.0, which matches
+// embedded-hal-async in renaming the trait to I2c.
 #[cfg(feature = "nb")]
+trait_set! {
+    /// Alias for i2c::Write (from embedded-hal < 1.0).
+    pub trait I2cWrite<E> = i2c::Write<Error = E>;
+    /// Alias for i2c::WriteRead (from embedded-hal < 1.0).
+    pub trait I2cWriteRead<E> = i2c::WriteRead<Error = E>;
+}
+
+#[cfg(feature = "async")]
+trait_set! {
+    /// Alias for i2c::I2c (from embedded-hal-async).
+    pub trait I2cWriteAsync<E> = i2c_async::I2c<Error = E>;
+    /// Alias for i2c::I2c (from embedded-hal-async).
+    pub trait I2cWriteReadAsync<E> = i2c_async::I2c<Error = E>;
+}
+
+#[maybe_async_cfg::maybe(sync(feature="nb", keep_self), async(feature="async", idents(I2cWrite)))]
 impl<I2C, E> Apds9960<I2C>
 where
-    I2C: i2c::Write<Error = E>,
+    I2C: I2cWrite<E, Error = E>,
 {
     /// Create new instance of the APDS9960 device.
     pub fn new(i2c: I2C) -> Self {
@@ -405,39 +428,8 @@ where
     }
 }
 
-#[cfg(feature = "async")]
-impl<I2C, E> Apds9960<I2C>
-where
-    I2C: async_i2c::I2c<Error = E>,
-{
-    /// Create new instance of the APDS9960 device.
-    pub fn new(i2c: I2C) -> Self {
-        Apds9960 {
-            i2c,
-            enable: register::Enable::default(),
-            config1: register::Config1::default(),
-            config2: register::Config2::default(),
-            gconfig1: register::GConfig1::default(),
-            gconfig4: register::GConfig4::default(),
-        }
-    }
-
-    /// Destroy driver instance, return I²C bus instance.
-    pub fn destroy(self) -> I2C {
-        self.i2c
-    }
-}
-
-#[cfg(feature = "async")]
-mod r#async;
-
-#[cfg(feature = "nb")]
 mod config;
-#[cfg(feature = "nb")]
 mod gesture;
-#[cfg(feature = "nb")]
 mod light;
-#[cfg(feature = "nb")]
 mod proximity;
-#[cfg(feature = "nb")]
 mod reading;
